refactor(tui): separate decision continuation owners
